+++ lib/overcommit/hook/pre_commit/travis_lint.rb
@@ -2,10 +2,10 @@ module Overcommit::Hook::PreCommit
   # Runs `travis-lint` against any modified Travis CI files.
   class TravisLint < Base
     def run
-      result = execute([executable] + applicable_files)
+      result = execute([executable, 'lint'] + applicable_files)
       return :pass if result.success?
 
-      [:fail, result.stdout.strip]
+      [:fail, (result.stdout + result.stderr).strip]
     end
   end
 end